kvserver: move replicaAppBatch to replica_app_batch.go
For #75729 we will want to work on replicaAppBatch and decompose it into
a new type appBatch around which replicaAppBatch can be layered.

As a first baby step, move appbatch-related code to its own
file to physically separate it from the `apply.StateMachine`.

This is 100% mechanical (using Goland's refactoring tools).

Release note: None
